requirements.txt
# Core ML & Data
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.2
lightgbm==4.5.0
imbalanced-learn==0.12.3
joblib==1.4.2
threadpoolctl==3.5.0

# Snowflake (versions compatibles)
snowflake-connector-python==3.10.0
snowflake-sqlalchemy==1.7.7
sqlalchemy==2.0.43
cryptography==42.0.

# MLflow & Tracking
mlflow==2.14.3
alembic==1.16.5
Flask==3.1.2
gunicorn==22.0.0    # ✅ compatible avec mlflow
graphene==3.4.3
scipy==1.13.1
matplotlib==3.9.2

# API & Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

# Dashboard
streamlit==1.38.0
plotly==5.24.1
altair==5.3.0

# Utils
python-dotenv==1.0.1
pyyaml==6.0.2
loguru==0.7.2
requests==2.32.3

# Testing
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.28.1
anyio==4.4.0
httpcore==1.0.5
iniconfig==2.0.0
pluggy==1.5.0
coverage==7.6.1
pygments==2.18.0

# Code Quality
black==24.8.0
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.3.6
pytokens==0.1.10
